[refactor] Create Modal Component
